requirements.txt
annotated-types==0.7.0
asgiref==3.11.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
contextlib2==21.6.0
cryptography==46.0.7
dj-database-url==3.1.2
email-validator==2.3.0
Django==6.0.4
django-allauth==65.16.1
django-cors-headers==4.9.0
django-ninja==1.6.2
django-ninja-extra==0.31.4
django-ninja-jwt==5.4.4
geographiclib==2.1
geopy==2.4.1
google-genai==1.3.0
google-cloud-monitoring==2.26.0
google-auth==2.49.2
google-auth-httplib2==0.3.1
google-auth-oauthlib==1.3.1
googlemaps==4.10.0
httplib2==0.31.2
idna==3.12
iniconfig==2.3.0
injector==0.24.0
oauthlib==3.3.1
packaging==26.0
pip==26.0.1
pluggy==1.6.0
psycopg2-binary==2.9.11
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydantic-settings==2.13.1
Pygments==2.20.0
PyJWT==2.12.1
pyparsing==3.3.2
pytest==9.0.3
coverage==7.6.1
pytest-cov==6.0.0
pytest-django==4.12.0
python-dotenv==1.2.2
requests==2.33.1
requests-oauthlib==2.0.0
sqlparse==0.5.5
typing_extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.6.3
